refactor :: remove the duplicate order list kept on user
Orders were stored in the orders collection and again as a DBRef list on the user,
so the same data had two owners and the list got loaded on every user fetch.
Removed the field and totalOrders now comes from a count on the orders collection,
which is cheap since userId is indexed.

+++ order_backend/src/main/java/com/reon/order_backend/dto/user/UserResponse.java
@@ -24,5 +24,5 @@ public class UserResponse {
     private LocalDateTime updatedOn;
 
     // only the count is sent here, actual orders are served by the order apis
-    private int totalOrders;
+    private long totalOrders;
 }
